requirements.txt
# Core Dependencies
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
implicit==0.7.2
sentence-transformers

# Deep Learning (for embeddings)
torch
torchvision
mysql-connector-python
# Database
psycopg2-binary==2.9.6
redis==4.6.0

# Message Queue
kafka-python==2.0.2

# API Framework
flask==2.3.2
flask-cors==4.0.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
schedule==1.2.0

# Logging & Monitoring
loguru==0.7.0

# Optional (for PDF export)
reportlab==4.0.4
matplotlib==3.7.2
seaborn==0.12.2
bs4
pyvi
faker